refactor install of virtual environment, repo and pip under the pi user on the first time starting up - fixes for docker files to stay in sync with this refactor in case non-root docker users become needed in the future

+++ install/pi/deploy-tools.sh
@@ -213,18 +213,17 @@ scipy_version="scipy-1.2.1-cp35-cp35m-linux_armv7l.whl"
 scipy_url="https://www.piwheels.org/simple/scipy/${scipy_version}#sha256=270be300233af556e6ee3f55a0ae237df0cb65ac85d47559010d7a9071f2e878"
 scipy_download_file="${DCMOUNTPATH}/opt/downloads/pip/${scipy_version}"
 
-if [[ ! -e ${scipy_download_file} ]]; then
+if [[ ! -e ${scipy_download_file} ]] && [[ -e /opt/stay_on_python3.5 ]]; then
     anmt "downloading newer scipy wheel file for install on the donkey car first time boot"
     wget ${scipy_url} -O ${scipy_download_file}
     chmod 666 ${scipy_download_file}
     chown ${DCUSER}:${DCUSER} ${scipy_download_file}
-fi
-
-if [[ ! -e ${scipy_download_file} ]]; then
-    err "failed to download ${scipy_version} from url: ${scipy_url}"
-    echo "wget ${scipy_url} -O ${scipy_download_file}"
-else
-    anmt "scipy ready for install on device at path: /opt/downloads/pip/${scipy_version}"
+    if [[ ! -e ${scipy_download_file} ]]; then
+        err "failed to download ${scipy_version} from url: ${scipy_url}"
+        echo "wget ${scipy_url} -O ${scipy_download_file}"
+    else
+        anmt "scipy ready for install on device at path: /opt/downloads/pip/${scipy_version}"
+    fi
 fi
 
 anmt "done installing tools"
